Move services into separate directory, update imports
