+++ Angular Frontend/src/app/toolbar/toolbar.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { LoginPageStateService } from '../login-page-state.service';
+import { LoginPageStateService } from '../../services/login-page-state.service';
 
 @Component({
   selector: 'app-toolbar',
