+++ Angular Frontend/src/services/login-page-state.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { LoginPageStateService } from './login-page-state.service';
+import { LoginPageStateService } from '../app/login-page-state.service';
 
 describe('LoginPageStateService', () => {
   let service: LoginPageStateService;
